Fix new file create bug.
The newFileName array was too small. The unit would attempt to find a new file name and then record a much larger file name into the array corrupting the RAM. This was regularly called when the unit was awaking between power cycles and seems to be the main culprit behind the lack of logging in issue #2. It doesn't seem to be related to exFat or variability between cards.