Allow the TX and RX pins to be used for the Terminal
This is a **big** change.
Via a new option in the Terminal Output menu, the TX pins and RX pins can now be used for the Terminal.
When this feature is enabled, all Serial data that would normally go just to Serial (USB) now goes to SerialLog too.
The menu choices functions accept data from Serial and SerialLog.
The SD card menu (option "s") can be accessed on either Serial (USB) or SerialLog (TX/RX). This menu will only appear on one or the other. ZMODEM support works on Serial (USB) or SerialLog (TX/RX) too.
This is very cool beacuase you can now hook up a Bluetooth Mate or BlueSMiRF or radio modem to the TX and RX pins and access the OLA via that instead!