adding back support for SN-GCJA5 particle sensor
The SN-GCJA5 particle sensor was on the same 0x33 address with MAX11615 ADC, however, the confidence of connection is low as it only performs basic isConnected(). Hence, in adding the support of MAX11615 ADC, it was removed to make sure the OLA recognize the ADC. Solution is to connect to MAX11615 ADC first with medium confidence (perform some init() ), tested and the OLA detect the MAX11615 ADC so can add back in the SN-GCJA5 particle sensor.